Fix circular dependency in tagging and connector modal
